+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 - PR #2679 AVRO reader: fix cuda errors when decoding compressed streams
 - PR #2651 Remove nvidia driver installation from ci/cpu/build.sh
 - PR #2697 Ensure csv reader sets datetime column time units
+- PR #2698 Return RangeIndex from contiguous slice of RangeIndex
 
 
 # cuDF 0.9.0 (Date TBD)